+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Tests for "1000000000D world".
 - Tests for "Rooks movements".
+- Tests for "Jack Silver: the casino".
 
 ## [3.4.0] - 2022-02-21
 ### Added