+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "1000000000D world".
 - Tests for "Rooks movements".
 - Tests for "Jack Silver: the casino".
+- Tests for "Fax machine".
 
 ## [3.4.0] - 2022-02-21
 ### Added