+++ CHANGELOG.md
@@ -4,12 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [3.5.0] - 2022-02-22
 ### Added
 - Tests for "1000000000D world".
 - Tests for "Rooks movements".
 - Tests for "Jack Silver: the casino".
 - Tests for "Fax machine".
+- Tests for "NGR - basic radar".
 
 ## [3.4.0] - 2022-02-21
 ### Added